% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-sketch-12-04-2018.R
\name{ars}
\alias{ars}
\title{Adaptive-Rejection Sampler Function}
\usage{
ars(n, g = NA, h = NA, h_prime = NA, D_left = -Inf,
  D_right = Inf, k = 10, step = 3, center = 0)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to perform a rejection sampling based on the 
theory developed by Gilks et al. (1992). This function employs the tangent
approach as opposed to the secant approach
}
\examples{
ars()

}
\keyword{cats}
