% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_penguin.R
\name{classify_penguin}
\alias{classify_penguin}
\title{Split a string}
\usage{
classify_penguin(bill_length, flipper_length)
}
\arguments{
\item{bill_length}{A character vector with one element.}

\item{flipper_length}{What to split on.}
}
\value{
A character vector.
}
\description{
Split a string
}
