% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{my_sample}
\alias{my_sample}
\title{Using two data.frames to create sample sets}
\usage{
my_sample(x, y)
}
\arguments{
\item{x}{a data.frame or matrix}

\item{y}{a data.frame or matrix}
}
\value{
2000 sample sets
}
\description{
Using two data.frames to create sample sets
}
\examples{
my_sample(GSE85589,GSE59856)#GSE85589 is a data.frame,GSE59856 is also a data.frame
}
