\name{pack-package}
\alias{pack-package}
\docType{package}
\title{
  Convert values to/from raw vectors
}
\description{
  \code{pack} allows R programmers to easily put their data into
  binary formats that other programs / machines can understand.
}
\author{
Author: Joshua M. Ulrich
Maintainer: Joshua M. Ulrich <josh.m.ulrich@gmail.com>
}
\references{
  \url{https://perldoc.perl.org/functions/pack}\cr
}
\keyword{ package }
