% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pack_circles}
\alias{pack_circles}
\title{Pack circles}
\usage{
pack_circles(
  polygon,
  radii,
  existing_circles,
  max_attempts = 2000L,
  seed = 1L,
  neat_edges = TRUE
)
}
\arguments{
\item{polygon}{A dataframe containing the x and y coordinates of vertices
a polygon into which circles will be packed}

\item{radii}{A numeric vector giving the radii of circles which will be
packed into the container specified in the polygon argument. This also
determines the maximum number of circles which the algorithm will attempt
to place.}

\item{existing_circles}{A dataframe of x and y coordinates and radii of any
existing circles to be avoided when packing the new ones. This allows
multiple iterations, for example to pack a shape within another shape.}

\item{max_attempts}{A scalar integer. How many times the algorithm will try
to place a circle in a new spot before giving up.}

\item{seed}{A scalar integer. Used to seed the random number generator.}

\item{neat_edges}{Logical. If true, no part of a circle can extend beyond
the container polygons boundary. If false, circles can extend beyond the
boundary as long as any part of the circle remains within the container.}
}
\description{
This function returns a dataframe of x and y coordinates and radii
of circles packed into an arbitrary polygon container.
}
