% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAtool.R
\name{QAtool}
\alias{QAtool}
\title{QAtool}
\usage{
QAtool(input, output, model = "general")
}
\arguments{
\item{input}{Can be .r .txt or .csv}

\item{output}{Is gonna be quite something}

\item{model}{'general' or 'personalised' Default is 'general'}
}
\description{
Applies a prebuilt model to Freesurfer output for QA
returns classification
}
\examples{
QAtool('FreesurferData.csv', 'QAdone.csv', model='general') )
QAtool()
}
