\name{pack-package}
\alias{pack-package}
\docType{package}
\title{
  Convert values to/from raw vectors
}
\description{
  \code{pack} allows R programmers to easily put their data into
  binary formats that other programs / machines can understand.
}
\details{
  \tabular{ll}{
    Package: \tab pack\cr
    Type: \tab Package\cr
    Version: \tab 0.1-1\cr
    Date: \tab 2008-08-22\cr
    License: \tab GPL-3\cr
    LazyLoad: \tab yes\cr
  }
}
\author{
Author: Josh Ulrich
Maintainer: Josh Ulrich <josh.m.ulrich@gmail.com>
}
\references{
  \url{http://perldoc.perl.org/functions/pack.html}\cr
}
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
