% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunc.R
\name{SolveSylvester}
\alias{SolveSylvester}
\title{SolveSylvester}
\usage{
SolveSylvester(A, B, C)
}
\arguments{
\item{A}{A square Matrix,}

\item{B}{A square Matrix.}

\item{C}{right hand side Matrix.}
}
\value{
A solve of \code{A} \code{B} \code{C}
}
\description{
Function used to solve Sylvester equation AX + XB = C
}
\examples{
SolveSylvester(matrix(c(1,0,-1,2,1,0,0,0,-1),3,3,byrow = TRUE),
matrix(c(2,1,0,0),2,2,byrow = TRUE),
matrix(c(2,2,1,1,1,1),3,2))

}
