% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunc.R
\name{SolveSylvester}
\alias{SolveSylvester}
\title{SolveSylvester}
\usage{
SolveSylvester(A, B, C)
}
\arguments{
\item{A}{A square Matrix or a real number,}

\item{B}{A square Matrix or a real number.}

\item{C}{A Matrix fits certain conditions.}
}
\value{
A solve of \code{AX + XB = C}, which is a Matrix.
}
\description{
Function used to solve Sylvester equation AX + XB = C.
This Package contains a function called \code{Solvesylvester()}
This function can be used to solve standard sylvester equations like AX + XB = C, where A, B, C are all square matrices with same dimensions.
Also, this function allows input matrices and real numbers, if real number is 0, it will degenerate to simple matrix equation.
Input can be \code{(Matrix, number, Matrix)} or \code{(number, Matrix, Matrix)} or \code{(Matrix, Matrix, Matrix)}, other input are not supported.
}
\examples{
SolveSylvester(matrix(c(1,0,-1,2,1,0,0,0,-1),3,3,byrow = TRUE),
matrix(c(2,1,0,0),2,2,byrow = TRUE),
matrix(c(2,2,1,1,1,1),3,2))

SolveSylvester(matrix(c(1,0,-1,2,1,0,0,0,-1),3,3,byrow = TRUE),
2,
matrix(c(2,2,1,1,1,1,3,4),3,3))
}
